@@ -4,22 +4,23 @@ version: 1.0.5
 homepage: https://www.yonomi.co
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.14.0 <3.0.0"
   
 dependencies:
-  http: ^0.12.2
-  yaml: ^2.2.1
-  dotenv: ^2.0.0
-  artemis: '>=6.0.0 <7.0.0'
-  json_annotation: ^3.1.0
   equatable: ^1.2.5
-  gql: '>=0.12.3 <1.0.0'
-  corsac_jwt: ^0.4.0
   meta: ^1.3.0
+  gql_exec: ^0.3.0
+  gql: ^0.13.0
+  gql_link: ^0.4.0
+  http: ^0.13.4
+  gql_http_link: ^0.4.0
+  yaml: ^3.1.0
+  dotenv: ^3.0.0
+  corsac_jwt: ^1.0.0-nullsafety.1
 
 dev_dependencies:
   test: ^1.15.7
   nock: ^1.1.2
-  build_runner: ^1.10.4
-  json_serializable: ^3.5.0
-  mockito: ^4.1.4
+  build_runner: ^2.1.4
+  gql_build: ^0.2.0
+
