@@ -1,19 +1,17 @@
 import 'package:test/test.dart';
-import 'package:yonomi_platform_sdk/graphql/devices/device_query.graphql.dart';
-import 'package:yonomi_platform_sdk/graphql/devices/thermostat/thermostat_queries.dart'
-    as ThermostatQueries;
 
-import 'package:yonomi_platform_sdk/repository/devices/devices_repository.dart';
-import 'package:yonomi_platform_sdk/repository/devices/lock_repository.dart';
-import 'package:yonomi_platform_sdk/repository/devices/thermostat_repository.dart';
-import 'package:yonomi_platform_sdk/request/request.dart' as yoRequest;
+import 'package:yonomi_platform_sdk/src/repository/devices/devices_repository.dart';
+import 'package:yonomi_platform_sdk/src/repository/devices/lock_repository.dart';
+import 'package:yonomi_platform_sdk/src/repository/devices/thermostat_repository.dart';
+import 'package:yonomi_platform_sdk/src/request/request.dart' as yoRequest;
+import 'package:yonomi_platform_sdk/third_party/yonomi_graphql_schema/schema.docs.schema.gql.dart';
 
 import '../../utils/test_fixtures.dart';
 
 void main() {
-  String testThermostatId, testLockId;
+  late String testThermostatId, testLockId;
 
-  yoRequest.Request request;
+  late yoRequest.Request request;
 
   setUpAll(() async {
     var tester = TestFixtures();
@@ -23,10 +21,6 @@ void main() {
     testThermostatId = await tester.getThermostatDeviceId(request);
 
     testLockId = await tester.getLockUnlockDeviceId(request);
-
-    assert(testThermostatId != null);
-
-    assert(testLockId != null);
   });
 
   test('getDevices returns device list for all traits', () async {
@@ -40,7 +34,11 @@ void main() {
     final device =
         await DevicesRepository.getDeviceDetails(request, testThermostatId);
     expect(device, isNotNull);
-    expect(device.traits.first.runtimeType, ThermostatTrait);
+    expect(
+        device.traits
+            .where((element) => element.runtimeType == ThermostatTrait)
+            .first is ThermostatTrait,
+        true);
   });
 
   test('getThermostatDetails gets thermostat details', () async {
@@ -55,15 +53,9 @@ void main() {
     expect(true, isTrue);
   });
 
-  test('setFanMode sets fanmode', () async {
-    await ThermostatRepository.setPointThermostat(
-        request, testThermostatId, 22);
-    expect(true, isTrue);
-  });
-
   test('setFanMode sets fan mode', () async {
     await ThermostatRepository.setMode(
-        request, testThermostatId, ThermostatQueries.ThermostatMode.heat);
+        request, testThermostatId, GThermostatMode.HEAT);
     expect(true, isTrue);
   });
 
@@ -72,7 +64,7 @@ void main() {
     final device =
         await DevicesRepository.getDeviceDetails(request, testLockId);
     expect(device, isNotNull);
-    expect(device.traits.first.runtimeType, LockUnlockTrait);
+    expect(device.traits.first.runtimeType, LockTrait);
   });
 
   test('getLockDetails gets lock details', () async {
@@ -85,146 +77,8 @@ void main() {
     expect(true, isTrue);
   });
 
-  test('responseToDeviceTraitConverter maps empty response to empty list',
-      () async {
-    List<DeviceDetailsMixin$DeviceTrait> responseTraits = [];
-
-    List<Trait> mappedTraits =
-        DevicesRepository.responseToDeviceTraitConverter(responseTraits);
-
-    expect(mappedTraits, isEmpty);
-  });
-
-  test(
-      'responseToDeviceTraitConverter maps single Lock DeviceTrait to LockUnlockTrait',
-      () async {
-    List<DeviceDetailsMixin$DeviceTrait> responseTraits = [
-      DeviceDetailsMixin$DeviceTrait.fromJson(
-        {
-          "__typename": "LockUnlockDeviceTrait",
-          "name": "LOCK_UNLOCK",
-          "properties": {"supportsIsJammed": true},
-          "state": {
-            "isLocked": {
-              "reported": {"value": false},
-            },
-          },
-        },
-      ),
-    ];
-
-    List<Trait> mappedTraits =
-        DevicesRepository.responseToDeviceTraitConverter(responseTraits);
-
-    expect(mappedTraits[0].runtimeType, LockUnlockTrait);
-  });
-
-  test(
-      'responseToDeviceTraitConverter maps single Thermostat DeviceTrait to ThermostatTrait',
-      () async {
-    List<DeviceDetailsMixin$DeviceTrait> responseTraits = [
-      DeviceDetailsMixin$DeviceTrait.fromJson(
-        {
-          "__typename": "ThermostatSettingDeviceTrait",
-          "name": "THERMOSTAT_SETTING",
-          "properties": {"supportsIsJammed": true},
-          "state": {
-            "targetTemperature": {
-              "reported": {"value": 22.0},
-            },
-          },
-        },
-      ),
-    ];
-
-    List<Trait> mappedTraits =
-        DevicesRepository.responseToDeviceTraitConverter(responseTraits);
-
-    expect(mappedTraits[0].runtimeType, ThermostatTrait);
-  });
-
-  test(
-      'responseToDeviceTraitConverter maps multiple responses to correct Trait objects',
-      () async {
-    List<DeviceDetailsMixin$DeviceTrait> responseTraits = [
-      DeviceDetailsMixin$DeviceTrait.fromJson(
-        {
-          "__typename": "LockUnlockDeviceTrait",
-          "name": "LOCK_UNLOCK",
-          "properties": {"supportsIsJammed": true},
-          "state": {
-            "isLocked": {
-              "reported": {"value": false},
-            },
-          },
-        },
-      ),
-      DeviceDetailsMixin$DeviceTrait.fromJson(
-        {
-          "__typename": "ThermostatSettingDeviceTrait",
-          "name": "THERMOSTAT_SETTING",
-          "properties": {"supportsIsJammed": true},
-          "state": {
-            "targetTemperature": {
-              "reported": {"value": 22.0},
-            },
-          },
-        },
-      ),
-    ];
-
-    List<Trait> mappedTraits =
-        DevicesRepository.responseToDeviceTraitConverter(responseTraits);
-
-    expect(mappedTraits[0].runtimeType, LockUnlockTrait);
-    expect(mappedTraits[1].runtimeType, ThermostatTrait);
-  });
-
-  test('responseToDeviceTraitConverter only maps Traits we support', () async {
-    List<DeviceDetailsMixin$DeviceTrait> responseTraits = [
-      DeviceDetailsMixin$DeviceTrait.fromJson(
-        {
-          "__typename": "LockUnlockDeviceTrait",
-          "name": "LOCK_UNLOCK",
-          "properties": {"supportsIsJammed": true},
-          "state": {
-            "isLocked": {
-              "reported": {"value": false},
-            },
-          },
-        },
-      ),
-      DeviceDetailsMixin$DeviceTrait.fromJson(
-        {
-          "__typename": "ThermostatSettingDeviceTrait",
-          "name": "THERMOSTAT_SETTING",
-          "properties": {"supportsIsJammed": true},
-          "state": {
-            "targetTemperature": {
-              "reported": {"value": 22.0},
-            },
-          },
-        },
-      ),
-      DeviceDetailsMixin$DeviceTrait.fromJson(
-        {
-          "__typename": "PowerDeviceTrait",
-          "name": "POWER",
-          "properties": {"supportsToggle": true},
-          "state": {
-            "power": {
-              "reported": {"value": true},
-            },
-          },
-        },
-      ),
-    ];
-
-    List<Trait> mappedTraits =
-        DevicesRepository.responseToDeviceTraitConverter(responseTraits);
-
-    expect(mappedTraits[0].runtimeType, LockUnlockTrait);
-    expect(mappedTraits[1].runtimeType, ThermostatTrait);
-    expect(mappedTraits.length, equals(2));
+  test('responseToDeviceTraitConverter maps empty response to empty list', () {
+    final convertedValue = DevicesRepository.responseToDeviceTraitConverter([]);
+    expect(convertedValue, isEmpty);
   });
 }
